feat: 체중 내역 저장 기능 구현
- 체중 내역 저장 기능 구현
- height, weight 데이터 타입 DECIMAL(4, 1)로 변경
- 평가 accept 시 더 낮은 무게로도 값이 바뀌는 문제 해결

+++ backend/bgs/src/main/java/com/ssafy/bgs/auth/dto/request/SocialSignupRequestDto.java
@@ -14,6 +14,6 @@ public class SocialSignupRequestDto {
     private String nickname;
     private LocalDate birthDate;
     private String sex;
-    private Integer height;
-    private Integer weight;
+    private Double height;
+    private Double weight;
 }
